+++ tweet_server/src/services/mod.rs
@@ -1,3 +1,4 @@
 pub mod features;
 pub mod ml_predict;
 pub mod recommend;
+pub mod queue;